% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ati-package.R
\docType{package}
\name{ati-package}
\alias{ati}
\alias{ati-package}
\title{ati: learning and teaching algorithmic trading and investment}
\description{
Package uses financial machine learning illustrate modern algorithmic trading and investment practices.
}
\author{
\strong{Maintainer}: Barry Quinn \email{b.quinn@qub.ac.uk} (\href{https://orcid.org/0000-0002-8637-9060}{ORCID})

}
\keyword{internal}
