% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{dir_create}
\alias{dir_create}
\alias{dir_remove}
\alias{dir_rename}
\alias{dir_copy}
\title{Create directory}
\usage{
dir_create(path, recursive = FALSE)

dir_remove(path, recursive = FALSE)

dir_rename(from, to)

dir_copy(from, to)
}
\arguments{
\item{path}{Path/name of directory to create}

\item{recursive}{Logical indicating whether to recursively create new
directory. This defaults to FALSE.}

\item{from}{Name of directory to rename/copy}

\item{to}{New directory}
}
\value{
Silently returns logical indicating success.

Logical indicating whether success
}
\description{
Creates a directory

Rename directory

Copy directory
}
