% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists.R
\name{dir_exists}
\alias{dir_exists}
\title{Directory exists}
\usage{
dir_exists(path)
}
\arguments{
\item{path}{Path}
}
\value{
Logical value with TRUE indicating it exists
}
\description{
Logical test of whether directory exists
}
