% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_info.R
\name{dir_name}
\alias{dir_name}
\title{Directory name}
\usage{
dir_name(path)
}
\arguments{
\item{path}{String representing path}
}
\value{
Name of containing
}
\description{
Returns directory name from a given path
}
\seealso{
Other file_info: \code{\link{file_info}},
  \code{\link{file_name}}
}
\concept{file_info}
