% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{dir_rename}
\alias{dir_rename}
\title{Directory rename}
\usage{
dir_rename(from, to)
}
\arguments{
\item{from}{Name of directory to rename}

\item{to}{New directory name}
}
\value{
Logical indicating whether success
}
\description{
Rename directory
}
