% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists.R
\name{file_exists}
\alias{file_exists}
\title{File exists}
\usage{
file_exists(path)
}
\arguments{
\item{path}{Path}
}
\value{
Logical value with TRUE indicating it exists
}
\description{
Logical test of whether file exists
}
