% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_info.R
\name{file_info}
\alias{file_info}
\title{File information}
\usage{
file_info(...)
}
\arguments{
\item{...}{Path names}
}
\value{
A data frame with path, size, isdir, permission, and timestamp
variables
}
\description{
Returns a data frame of file information
}
\seealso{
Other file_info: \code{\link{dir_name}},
  \code{\link{file_name}}
}
\concept{file_info}
