% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_info.R
\name{file_name}
\alias{file_name}
\title{File name}
\usage{
file_name(path)
}
\arguments{
\item{path}{Path}
}
\value{
Name of file
}
\description{
Returns file name from a given path
}
\seealso{
Other file_info: \code{\link{dir_name}},
  \code{\link{file_info}}
}
\concept{file_info}
