% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.R
\name{file_path}
\alias{file_path}
\alias{\%FP\%}
\alias{fp}
\title{File path}
\usage{
file_path(...)

lhs \%FP\% rhs

fp(...)
}
\arguments{
\item{...}{Names to path}

\item{lhs}{Left hand side}

\item{rhs}{Right hand side}
}
\value{
Constructed/pasted together path name
}
\description{
Builds a file path
}
