% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{file_tmp}
\alias{file_tmp}
\title{Temporary file}
\usage{
file_tmp(dir = NULL, ext = NULL)
}
\arguments{
\item{dir}{Optional name of temporary directory}

\item{ext}{Optional extension to use}
}
\value{
Name of temporary file
}
\description{
Create temporary file
}
