% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_file.R
\name{find_file}
\alias{find_file}
\title{Find file}
\usage{
find_file(...)
}
\arguments{
\item{...}{Name of file (can include containing directory to narrow search)}
}
\value{
Path if found. NULL if not.
}
\description{
Looks for file in common spots and returns full path
}
