% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fml-package.R
\docType{package}
\name{fml-package}
\alias{fml}
\alias{fml-package}
\title{fml: File Management and Location Tools}
\description{
The fml package provides a streamlined and simplified interface for
interacting with files. It serves as a wrapper around numerous base
functions e.g., \link[base:path.expand]{base::path.expand}, \link[base:file.path]{base::file.path}, \link[base:list.files]{base::list.files},
\link[base:list.dirs]{base::list.dirs}, etc.
}
\section{file functions}{

\strong{File}s are non-directory, content endpoints (docs, text, html, script, data, etc.)
}

\section{dir functions}{

\strong{Dir}ectories are non-file, container endpoints (folders that can contain
files and/or subfolders)
}

\section{path functions}{

\strong{Path}s are file or directory endpoints (everything is a path)
}

\seealso{
Useful links:
\itemize{
  \item \url{https://fml.mikewk.com}
  \item Report bugs at \url{https://github.com/mkearney/fml/issues}
}

}
\author{
\strong{Maintainer}: Michael W. Kearney \email{kearneymw@missouri.edu} (0000-0002-0730-4694)

}
