% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.R
\name{fp}
\alias{fp}
\alias{\%FP\%}
\alias{path_path}
\alias{file_path}
\title{File path}
\usage{
fp(...)

... \%FP\% NA

path_path(...)

file_path(...)
}
\arguments{
\item{...}{Names to path}
}
\value{
String of path name
}
\description{
Builds a file path
}
