% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRndCov.R
\name{getRndCoV}
\alias{getRndCoV}
\title{Random Matrix with Signal
Function to create a random matrix with a number of ground truth facts or signals.}
\usage{
getRndCoV(nCols, nFacts)
}
\arguments{
\item{nCols}{number of columns or stocks}

\item{nFacts}{number of signals to add to the random matrix}
}
\value{
matrix
}
\description{
Random Matrix with Signal
Function to create a random matrix with a number of ground truth facts or signals.
}
\examples{
nCols=1000
nFacts=100
alpha<-q<-0.95
rtns=array(rnorm(nCols*q*nCols),dim = c(nCols*q,nCols))
cov1=cov(rtns)
cov1=alpha*cov1+(1-alpha)*getRndCoV(nCols = nCols,nFacts = nFacts)
corr0=cov2cor(cov1)
eigs=eigen(corr0)
}
\references{
López de Prado, Marcos. 2020. “Machine Learning for Asset Managers.” In Elements in Quantitative Finance. Cambridge University Press.
}
