% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here.R
\name{here}
\alias{here}
\title{Where is here}
\usage{
here(...)
}
\arguments{
\item{...}{Parts comprising a single path}
}
\value{
Returns the supplied file name appended onto the root directory
}
\description{
Build a path from an R project root
}
\details{
Looks for a .Rproj or .here file and then appends the supplied file
}
\seealso{
Other locate: \code{\link{find_file}},
  \code{\link{search_file}}
}
\concept{locate}
