% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{List files}
\usage{
list_files(path = ".", full = TRUE, all = FALSE, recursive = FALSE,
  pattern = NULL, ignore.case = FALSE, invert = FALSE)
}
\arguments{
\item{path}{Directory from which to locate dirs–defaults to current working
directory.}

\item{full}{Logical indicating whether to return full paths. Defaults
to TRUE.}

\item{all}{Logical indicating whether to return all (include dot files)
file names. Defaults to FALSE.}

\item{recursive}{Logical indicating whether to recursively list files.}

\item{pattern}{Regular expression pattern on which to filter (return) matching
results. Defaults to NULL, meaning all qualifying files will be returned.}

\item{ignore.case}{Logical indicating whether to ignore case if
\code{pattern} is supplied–defaults to FALSE.}

\item{invert}{Logical indicating whether to invert (return those that DO NOT
match) regex matching results–defaults to FALSE.}
}
\value{
A character vector of directories
}
\description{
Returns names of files located in a given directory. It's similar to
list.files() only with better defaults and behaviors.
}
\examples{

## list files in working directory
list_files()

}
