% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_files}
\alias{match_files}
\title{Match files}
\usage{
match_files(pat)
}
\arguments{
\item{pat}{File pattern on which to match where * means 0 or more, ? means
once, and \[ indicates the start of a string.}
}
\value{
Matching files
}
\description{
List files matching search operator-assisting patterns
}
