% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_pdf.R
\name{mp_pdf}
\alias{mp_pdf}
\title{Marcenko Pastur pdf function}
\usage{
mp_pdf(var = 1, t, m, pts)
}
\arguments{
\item{var}{variance}

\item{t}{time}

\item{m}{number of stocks}

\item{pts}{number of data points to sample}
}
\description{
create m-p distribution for random matrix
}
\keyword{Marcenko-Pastur}
