% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numBins}
\alias{numBins}
\title{Discretisation of a continuous variable}
\usage{
numBins(nObs, corr = NULL)
}
\arguments{
\item{nObs}{number of observations}

\item{corr}{the correlation for the bivariate case}
}
\description{
Discretisation of a continuous variable
}
