% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.R
\name{package_file}
\alias{package_file}
\title{Package file}
\usage{
package_file(path, package = "base")
}
\arguments{
\item{path}{Path}

\item{package}{Name of package in which file exists.}
}
\value{
Full path to package
}
\description{
Construct path to an R package file
}
