% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.R
\name{path_expand}
\alias{path_expand}
\title{Expand path}
\usage{
path_expand(path = ".")
}
\arguments{
\item{path}{Character vector of paths to normalize}
}
\value{
Full/expanded canonical paths
}
\description{
Expands path to canonical form for the platform
}
