% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists.R
\name{read_access}
\alias{read_access}
\title{Check for read access}
\usage{
read_access(path)
}
\arguments{
\item{path}{Path}
}
\value{
Logical indicating wether read permission exists
}
\description{
Check for read access
}
