% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_files.R
\name{search_file}
\alias{search_file}
\title{Search file}
\usage{
search_file(...)
}
\arguments{
\item{...Name}{to match (can include containing directory to narrow search)}
}
\value{
Path if found. NULL if not.
}
\description{
Searches for matching file names in common spots and returns full path
}
