% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists.R
\name{write_access}
\alias{write_access}
\title{Check for write access}
\usage{
write_access(path)
}
\arguments{
\item{path}{Path}
}
\value{
Logical indicating wether write permission exists
}
\description{
Check for write access
}
