% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{dir_create}
\alias{dir_create}
\alias{dir_remove}
\title{Create directory}
\usage{
dir_create(path, recursive = FALSE)

dir_remove(path, recursive = FALSE)
}
\arguments{
\item{path}{Path/name of directory to create}

\item{recursive}{Logical indicating whether to recursively create new
directory. This defaults to FALSE.}

\item{recursive}{Logical indicating whether to recursively remove files.}
}
\value{
Silently returns logical indicating success.
}
\description{
Creates a directory
}
