% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists.R
\name{dir_exists}
\alias{dir_exists}
\title{Directory exists}
\usage{
dir_exists(...)
}
\arguments{
\item{...}{Enter parts to dir path.}
}
\value{
Logical value
}
\description{
Logical test of whether directory exists
}
