% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R, R/exists.R
\name{file_create}
\alias{file_create}
\alias{file_remove}
\alias{file_rename}
\alias{file_copy}
\alias{file_exists}
\alias{symlink_exists}
\title{Create file}
\usage{
file_create(path, recursive = FALSE)

file_remove(path)

file_rename(from, to)

file_copy(from, to, overwrite = FALSE, recursive = FALSE)

file_exists(path)

symlink_exists(path)
}
\arguments{
\item{path}{Path}

\item{recursive}{Logical indicating whether to recursively create new
file (directories created as necessary). This defaults to FALSE.}

\item{from}{File to rename/copy}

\item{to}{New/copy name}

\item{overwrite}{Logical indicating whether to overwrite}
}
\value{
Silently returns logical indicating success.

Logical value with TRUE indicating it exists
}
\description{
Creates a file

Rename file

copy file

Logical test of whether file exists

Check whether a symlink exists
}
\seealso{
Other exists: \code{\link{dir_create}}

Other exists: \code{\link{dir_create}}
}
\concept{exists}
