% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exists.R
\name{file_exists}
\alias{file_exists}
\title{File exists}
\usage{
file_exists(...)
}
\arguments{
\item{...}{Enter parts to file path.}
}
\value{
Logical value
}
\description{
Logical test of whether file exists
}
