% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistExpectedMaxSR.R
\name{getDistExpectedMaxSR}
\alias{getDistExpectedMaxSR}
\title{Maximum Sharpe Ratio Monte Carlo Simulator}
\usage{
getDistExpectedMaxSR(nSims, nTrails, meanSR, stdSR)
}
\arguments{
\item{nSims}{number of simulations per trail}

\item{nTrails}{a numeric vector of trails starting at 2}

\item{meanSR}{mean value of Sharpe ratio under the null of a false positive strategy}

\item{stdSR}{standard deviation of Sharpe ratio under the null of a false positive strategy}
}
\value{
tibble
}
\description{
Simulates the distribution of the expected Maximum Sharpe ratio for a set number of trails
}
