% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpectedMaxSR.R
\name{getExpectedMaxSR}
\alias{getExpectedMaxSR}
\title{Expected Maximum Sharpe Ratio}
\usage{
getExpectedMaxSR(nTrails, meanSR, stdSR)
}
\arguments{
\item{stdSR}{}
}
\value{

}
\description{
Calculcate the theoretical Maximum Sharpe Ratio according to the False Discovery theorem first proposed by Bailey et al. 2014.
}
\references{
Bailey, D., J. Borwein, M. López de Prado, and J. Zhu 2014: “Pseudo- mathematics and financial charlatanism: The effects of backtest overfitting on out-of-sample performance.” Notices of the American Mathematical Society, Vol. 61, No. 5, pp. 458–471. Available at http://ssrn.com/abstract=2308659
}
