% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_file.R
\name{list_dirs}
\alias{list_dirs}
\title{List directories}
\usage{
list_dirs(path = ".", full.names = TRUE, all.names = FALSE)
}
\arguments{
\item{path}{Directory from which to locate dirs, defaults to current working
directory.}

\item{all.names}{Logical indicating whether to return all (include dot files)
dir names. Defaults to FALSE.}

\item{fulle.names}{Logical indicating whether to return full paths. Defaults
to TRUE.}
}
\value{
A character vector of directories
}
\description{
Returns names of directories located in a given directory. It's similar to
list.dirs() only with better defaults and behavior.
}
