% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_files.R
\name{search_file}
\alias{search_file}
\title{Search file}
\usage{
search_file(pat, dir = get_path("."))
}
\arguments{
\item{pat}{Name to match (can include containing directory to narrow search)}

\item{dir}{Optional, directory from which to start/limit search (only
subfolders will be searched)}
}
\value{
Path if found. NULL if not.
}
\description{
Searches for matching file names in common spots and returns full path
}
\seealso{
Other locate: \code{\link{find_file}}, \code{\link{here}}
}
\concept{locate}
